lint

+++ rust/worker/src/segment/metadata_segment.rs
@@ -1388,7 +1388,7 @@ mod test {
             .await
             .expect("Commit for segment writer failed");
         let total_keys = record_flusher.total_keys();
-        assert_eq!(total_keys, Some(2 as u64));
+        assert_eq!(total_keys, Some(2_u64));
         let metadata_flusher = metadata_writer
             .commit()
             .await
